made scripts runnable in py3 too

+++ screen_shot_cap.py
@@ -22,31 +22,31 @@ numFaces = 0
 font = cv2.FONT_HERSHEY_SIMPLEX
 
 def detectFaces(gray,scaleFactor = 1.1,minNeighbors=5,minSize=(30,30)):
-	faces = face_cascade.detectMultiScale(gray, scaleFactor=scaleFactor,
+    faces = face_cascade.detectMultiScale(gray, scaleFactor=scaleFactor,
         minNeighbors=minNeighbors,
         minSize=minSize)
-	return faces
+    return faces
 
 
 def saveFaceImg(x,y,w,h,numFaces):
-	margin = max(w,h)/2
-	cropFace = img[y-margin:y+h+margin,x-margin:x+w+margin]
-	saveFName = newpath+'roi'+str(numFaces)+'.png'
-	cv2.imwrite(saveFName,cropFace)
-	print ('saved face img',numFaces)
-	return cropFace,saveFName
+    margin = int(max(w,h)/2)
+    cropFace = img[y-margin:y+h+margin,x-margin:x+w+margin]
+    saveFName = newpath+'roi'+str(numFaces)+'.png'
+    cv2.imwrite(saveFName,cropFace)
+    print ('saved face img',numFaces)
+    return cropFace,saveFName
 
 def estSex(saveFName):
-	# resized_img = resize(cropFace)
-	# arr = im2Array(resized_img)
-	arr = np.array([loadImg2Array(saveFName)])
-	sex = predict_vgg(model,arr)
-	return sex
+    # resized_img = resize(cropFace)
+    # arr = im2Array(resized_img)
+    arr = np.array([loadImg2Array(saveFName)])
+    sex = predict_vgg(model,arr)
+    return sex
 
 def labelFaces(x,y,w,h):
-	cv2.rectangle(img,(x,y),(x+w,y+h),(255,0,0),2)
-	imgLabel = sex
-	cv2.putText(img,imgLabel,(x, y), font, 1,(0,255,0),1,cv2.LINE_AA)
+    cv2.rectangle(img,(x,y),(x+w,y+h),(255,0,0),2)
+    imgLabel = sex
+    cv2.putText(img,imgLabel,(x, y), font, 1,(0,255,0),1,cv2.LINE_AA)
 
 model = loadVGG()
 
@@ -64,16 +64,16 @@ while(True):
     faces = detectFaces(gray)
 
     for (x,y,w,h) in faces:
-		numFaces += 1
-		cropFace,saveFName = saveFaceImg(x,y,w,h,numFaces)
+        numFaces += 1
+        cropFace,saveFName = saveFaceImg(x,y,w,h,numFaces)
 
-		if (os.stat(saveFName).st_size) > 0:
-			i_w,i_h = Image.open(saveFName).size
-			if 1.2 >i_w/i_h > 0.8:
-				sex = estSex(saveFName)
-				labelFaces(x,y,w,h)
-			else:
-				os.remove(saveFName)
+        if (os.stat(saveFName).st_size) > 0:
+            i_w,i_h = Image.open(saveFName).size
+            if 1.2 >i_w/i_h > 0.8:
+                sex = estSex(saveFName)
+                labelFaces(x,y,w,h)
+            else:
+                os.remove(saveFName)
 
     screenshot = cv2.resize(img, (0,0), fx=0.5, fy=0.5) 
     cv2.imshow("screenshot", screenshot)
@@ -82,5 +82,5 @@ while(True):
 
 
     if cv2.waitKey(1) & 0xFF == ord('q'):
-    	break
+        break
 cv2.destroyAllWindows()
